requirements.txt
uvicorn==0.28.0
python-dotenv==1.0.0
pydantic==2.3.0
requests==2.31.0
fastapi==0.109.2
pyodbc
gunicorn
python-jose[cryptography]
bcrypt